* Implement #197

+++ Source/Krypton Components/Krypton.Ribbon/Palette/PaletteGalleryBackBorder.cs
@@ -244,7 +244,7 @@ namespace Krypton.Ribbon
         /// </summary>
         /// <param name="state">Palette value should be applicable to this state.</param>
         /// <returns>Border rounding.</returns>
-        public int GetBorderRounding(PaletteState state)
+        public float GetBorderRounding(PaletteState state)
         {
             return 0;
         }
