+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteRedirect/PaletteRedirectBorder.cs
@@ -265,8 +265,8 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="style">Border style.</param>
         /// <param name="state">Palette value should be applicable to this state.</param>
-        /// <returns>Integer rounding.</returns>
-        public override int GetBorderRounding(PaletteBorderStyle style, PaletteState state)
+        /// <returns>Float rounding.</returns>
+        public override float GetBorderRounding(PaletteBorderStyle style, PaletteState state)
         {
             IPaletteBorder inherit = GetInherit(state);
 
@@ -310,7 +310,7 @@ namespace Krypton.Toolkit
             IPaletteBorder inherit = GetInherit(state);
 
             return inherit?.GetBorderImageAlign(state) ?? Target.GetBorderImageAlign(style, state);
-        }    
+        }
         #endregion
 
         #region Implementation
