+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteRedirect/PaletteRedirectDouble.cs
@@ -558,8 +558,8 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="style">Border style.</param>
         /// <param name="state">Palette value should be applicable to this state.</param>
-        /// <returns>Integer rounding.</returns>
-        public override int GetBorderRounding(PaletteBorderStyle style, PaletteState state)
+        /// <returns>Float rounding.</returns>
+        public override float GetBorderRounding(PaletteBorderStyle style, PaletteState state)
         {
             IPaletteDouble inherit = GetInherit(state);
